moved StartMastodon and added gradle shadow build option
